chore(frontend): remove WalletConnect support
WalletConnectContent renamed to WalletContent, without the WalletConnect wiring
