+++ frontend/package.json
@@ -46,7 +46,6 @@
     "motion": "^13.1.0",
     "next": "16.1.6",
     "next-themes": "^0.4.6",
-    "qrcode": "^1.5.4",
     "radix-ui": "^1.6.7",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
@@ -67,7 +66,6 @@
     "@types/d3": "^7.4.3",
     "@types/jest": "^30.0.0",
     "@types/node": "^22.19.9",
-    "@types/qrcode": "^1.5.6",
     "@types/react": "^19.2.13",
     "@types/react-dom": "^19.2.3",
     "baseline-browser-mapping": "^2.9.19",
