--- frontend/src/components/connector-kit/common/WalletConnectIcon.tsx
@@ -1,10 +1,0 @@
-export function WalletConnectIcon({ className }: { className?: string }) {
-  return (
-    <svg className={className} viewBox="0 0 480 480" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path
-        d="M126.613 168.813C193.707 101.719 302.293 101.719 369.387 168.813L377.507 176.933C380.853 180.28 380.853 185.707 377.507 189.053L350.453 216.107C348.787 217.773 346.067 217.773 344.4 216.107L333.387 205.093C287.2 158.907 213.293 158.907 167.107 205.093L155.2 216.987C153.533 218.653 150.813 218.653 149.147 216.987L122.093 189.933C118.747 186.587 118.747 181.16 122.093 177.813L126.613 168.813ZM426.667 225.88L450.827 250.04C454.173 253.387 454.173 258.813 450.827 262.16L343.44 369.547C340.093 372.893 334.667 372.893 331.32 369.547L255.427 293.653C254.587 292.813 253.227 292.813 252.387 293.653L176.493 369.547C173.147 372.893 167.72 372.893 164.373 369.547L56.9867 262.16C53.64 258.813 53.64 253.387 56.9867 250.04L81.1467 225.88C84.4933 222.533 89.92 222.533 93.2667 225.88L169.16 301.773C170 302.613 171.36 302.613 172.2 301.773L248.093 225.88C251.44 222.533 256.867 222.533 260.213 225.88L336.107 301.773C336.947 302.613 338.307 302.613 339.147 301.773L415.04 225.88C418.387 222.533 423.813 222.533 426.667 225.88Z"
-        fill="#3B99FC"
-      />
-    </svg>
-  );
-}
